DbConnectionFactory to IDisposable
